Move temporary types in data plugin to common/types. (#50280)
* Move temporary ui imported types in data plugin to common/types.
Remove unnecessary mocks.

* Updated dashboard, discover and visualize mappings - type removal

* Revert "Updated dashboard, discover and visualize mappings - type removal"

This reverts commit 5e931ce62a070eb9d4ea254808e75cc3fd47833d.
