+++ src/plugins/data/common/es_query/filters/phrase_filter.ts
@@ -19,7 +19,7 @@
 
 import { get, isPlainObject } from 'lodash';
 import { Filter, FilterMeta } from './meta_filter';
-import { IndexPattern, Field } from './types';
+import { IndexPattern, Field } from '../../types';
 
 export type PhraseFilterMeta = FilterMeta & {
   params?: {
