+++ src/plugins/data/common/types.ts
@@ -27,3 +27,4 @@ export * from './kbn_field_types/types';
 // will be replaces after Fieds / IndexPattern will be moved into new platform
 export type Field = any;
 export type IndexPattern = any;
+export type StaticIndexPattern = any;
