+++ src/plugins/data/public/suggestions_provider/types.ts
@@ -16,8 +16,6 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-// Should be import { Field } from './index_patterns';
-export type Field = any;
+import { Field } from '..';
 
 export type IGetSuggestions = (index: string, field: Field, query: string, boolFilter?: any) => any;
