+++ src/plugins/data/public/suggestions_provider/value_suggestions.test.ts
@@ -19,7 +19,6 @@
 
 // TODO: remove when index patterns are moved here.
 jest.mock('ui/new_platform');
-jest.mock('ui/index_patterns');
 
 import { stubIndexPattern, stubFields } from '../stubs';
 import { getSuggestionsProvider } from './value_suggestions';
